fix: rename sw service
