+++ apps/sw/src/health/health.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+
+import { SwHealthController } from '@apps/sw/health/interface/health.controller';
+
+@Module({
+  imports: [],
+  controllers: [SwHealthController],
+  providers: [],
+})
+export class HealtyModule {}
