+++ apps/sw/src/sw.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 
-import { HealtyModule } from '@apps/swapi/health/health.module';
+import { HealtyModule } from '@apps/sw/health/health.module';
 
 @Module({
   imports: [HealtyModule],
   controllers: [],
   providers: [],
 })
-export class SwapiModule {}
+export class SwModule {}
