+++ apps/sw/test/jest-e2e.json
