--- apps/swapi/src/health/health.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { SwapiHealtyController } from '@apps/swapi/health/interface/health.controller';
-
-@Module({
-  imports: [],
-  controllers: [SwapiHealtyController],
-  providers: [],
-})
-export class HealtyModule {}
