--- apps/swapi/src/health/interface/health.controller.spec.ts
@@ -1,23 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-
-import { SwapiHealtyController } from '@apps/swapi/health/interface/health.controller';
-
-describe('SwapiHealtyController', () => {
-  let swapiHealtyController: SwapiHealtyController;
-
-  beforeEach(async () => {
-    const app: TestingModule = await Test.createTestingModule({
-      controllers: [SwapiHealtyController],
-    }).compile();
-
-    swapiHealtyController = app.get<SwapiHealtyController>(
-      SwapiHealtyController,
-    );
-  });
-
-  describe('root', () => {
-    it('should return "Hello World!"', () => {
-      expect(swapiHealtyController.health()).toBe('health');
-    });
-  });
-});
