+++ package.json
@@ -7,14 +7,14 @@
   "license": "UNLICENSED",
   "scripts": {
     "prebuild": "rimraf dist",
-    "build": "npm run build:swapi && npm run build:gateway",
+    "build": "npm run build:sw && npm run build:gateway",
     "build:gateway": "nest build gateway",
-    "build:swapi": "nest build swapi",
+    "build:sw": "nest build sw",
     "format": "prettier --write \"apps/**/*.ts\" \"libs/**/*.ts\"",
     "start": "nest start",
-    "start:dev": "concurrently 'npm run start:dev:swapi' 'npm run start:dev:gateway'",
+    "start:dev": "concurrently 'npm run start:dev:sw' 'npm run start:dev:gateway'",
     "start:dev:gateway": "nest start gateway --watch",
-    "start:dev:swapi": "nest start swapi --watch",
+    "start:dev:sw": "nest start sw --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/apps/gateway/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
