+++ tsconfig.json
@@ -18,7 +18,7 @@
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
     "paths": {
-      "@apps/swapi/*": ["apps/swapi/src/*"]
+      "@apps/sw/*": ["apps/sw/src/*"]
     }
   }
 }
